# doctolib/swagger/swaggerDoctolib.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Doctolib Management",
        "description": "Doctolib Management",
        "version": "V1"
    },
    "paths": {
...
        "/doctors": {
            "get": {
                "tags": [
                    "doctors"
                ],
                "summary": "Returns a map of all doctor or a map of doctor according to a designated specialty",
                "description": "Returns a map of all doctor or a map of doctor according to a designated specialty",
                "operationId": "getDoctors",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DoctorDTO"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Please contact us"
                    },
                    "404": {
                        "description": "No doctors found"
                    }
                }
            },
            "post": {
                "tags": [
                    "doctors"
                ],
                "summary": "Create a doctor",
                "description": "Create a doctor",
                "operationId": "createDoctor",
                "responses": {
                    "201": {
                        "description": "Successful operation but no content"
                    },
                    "500": {
                        "description": "Please contact us"
                    }
                }
            }
        },
        "/doctors/{id}": {
            "put": {
                "tags": [
                    "doctors"
                ],
                "summary": "Update a doctor according to ID",
                "description": "Update a doctor according to ID",
                "operationId": "updateDoctor",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the doctor that needs to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DoctorDTO"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Please contact us"
                    }
                }
            },
            "delete": {
                "tags": [
                    "doctors"
                ],
                "summary": "Delete a doctor according to ID",
                "description": "Delete a doctor according to ID",
                "operationId": "deleteDoctor",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the doctor that needs to be deleted",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation but no content"
                    },
                    "500": {
                        "description": "Please contact us"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DoctorDTO": {
                "title": "DoctorDTO model",
                "description": "DoctorDTO model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "integer"
                    },
                    "email": {
                        "title": "email",
                        "description": "email",
                        "type": "string",
                        "format": "string"
                    },
                    "password": {
                        "title": "password",
                        "description": "password",
                        "type": "string",
                        "format": "string"
                    },
                    "lastName": {
                        "title": "lastName",
                        "description": "lastName",
                        "type": "string",
                        "format": "string"
                    },
                    "firstName": {
                        "title": "firstName",
                        "description": "firstName",
                        "type": "string",
                        "format": "string"
                    },
                    "specialty": {
                        "title": "specialty",
                        "description": "specialty",
                        "type": "string",
                        "format": "string"
                    },
                    "addressNum": {
                        "title": "addressNum",
                        "description": "addressNum",
                        "type": "integer",
                        "format": "integer"
                    },
                    "addressStreet": {
                        "title": "addressStreet",
                        "description": "addressStreet",
                        "type": "string",
                        "format": "string"
                    },
                    "postalCodeAddress": {
                        "title": "postalCodeAddress",
                        "description": "postalCodeAddress",
                        "type": "integer",
                        "format": "integer"
                    },
                    "townAddress": {
                        "title": "townAddress",
                        "description": "townAddress",
                        "type": "string",
                        "format": "string"
                    }
                },
                "type": "object"
            },
...
        }
    }
}